# docs/swagger.yaml
basePath: /
definitions:
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.ApiMessage:
    properties:
      message:
        type: string
    type: object
...
  models.ApiShowSubtasks:
    properties:
      subtasks:
        items:
          $ref: '#/definitions/models.SubtasksData'
        type: array
    type: object
  models.ApiShowTasks:
    properties:
      tasks:
        items:
          $ref: '#/definitions/models.TasksData'
        type: array
    type: object
  models.ApiSubtaskData:
    properties:
      comment:
        example: Sugar-free
        type: string
      name:
        example: Coca-Cola
        type: string
      task_id:
        example: 1023456789
        type: integer
    type: object
  models.ApiTaskData:
    properties:
      comment:
        example: Go to the supermarket on the way home
        type: string
      list_id:
        example: 1023456789
        type: integer
      name:
        example: Buy drinks
        type: string
    type: object
...
  models.SubtaskEditData:
    properties:
      categories:
        example:
        - Party
        - Shoping
        - Today
        items:
          type: string
        type: array
      comment:
        example: Sugar-free
        type: string
      done:
        example: true
        type: boolean
      end_time:
        example: 2077-12-10 13:13
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 1
        type: integer
      name:
        example: Pepsi
        type: string
      special:
        example: true
        type: boolean
    type: object
  models.SubtasksData:
    properties:
      categories:
        example:
        - Party
        - Shoping
        items:
          type: string
        type: array
      comment:
        example: Sugar-free
        type: string
      done:
        type: boolean
      end_time:
        example: 2077-12-10 13:13
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 0
        type: integer
      name:
        example: Coca-Cola
        type: string
      special:
        type: boolean
    type: object
  models.TaskEditData:
    properties:
      categories:
        example:
        - Party
        - Shoping
        - Today
        items:
          type: string
        type: array
      comment:
        example: Go to the supermarket on the way home
        type: string
      done:
        example: true
        type: boolean
      end_time:
        example: 2077-12-10 13:13
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 1
        type: integer
      name:
        example: Buy new drinks
        type: string
      special:
        example: true
        type: boolean
    type: object
  models.TasksData:
    properties:
      categories:
        example:
        - Party
        - Shoping
        items:
          type: string
        type: array
      comment:
        example: Go to the supermarket on the way home
        type: string
      done:
        type: boolean
      end_time:
        example: 2077-12-10 13:13
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 0
        type: integer
      name:
        example: Buy drinks
        type: string
      special:
        type: boolean
    type: object
...
info:
  contact:
    email: nktkln@nktkln.com
    name: API Support
    url: https://nktkln.com
  description: some description
  license:
    name: MIT
    url: http://github.com/
  title: ToDo API
  version: "1.0"
paths:
...
  /todo/subtask/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Subtask data
        in: body
        name: SubtaskData
        required: true
        schema:
          $ref: '#/definitions/models.ApiSubtaskData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Create subtask
      tags:
      - Working with subtasks
  /todo/subtask/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: The id of the subtask to be deleted
        in: query
        name: subtask_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Delete subtask
      tags:
      - Working with subtasks
  /todo/subtask/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Subtask data
        in: body
        name: SubtaskData
        required: true
        schema:
          $ref: '#/definitions/models.SubtaskEditData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Edit subtask
      tags:
      - Working with subtasks
  /todo/subtask/show:
    get:
      consumes:
      - application/json
      parameters:
      - description: Task id with subtasks
        in: query
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiShowSubtasks'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Shows all subtasks in the task
      tags:
      - Working with subtasks
  /todo/task/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task data
        in: body
        name: TaskData
        required: true
        schema:
          $ref: '#/definitions/models.ApiTaskData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Create task
      tags:
      - Working with tasks
  /todo/task/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: The id of the task to be deleted
        in: query
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Delete task
      tags:
      - Working with tasks
  /todo/task/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Task data
        in: body
        name: TaskData
        required: true
        schema:
          $ref: '#/definitions/models.TaskEditData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Edit task
      tags:
      - Working with tasks
  /todo/task/show:
    get:
      consumes:
      - application/json
      parameters:
      - description: List id with tasks
        in: query
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiShowTasks'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Shows all tasks in the list
      tags:
      - Working with tasks
...
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"